+++ app/DynamicsAdapter/DynamicsAdapter.Web/PersonSearch/PersonSearchController.cs
@@ -24,14 +24,14 @@ namespace DynamicsAdapter.Web.PersonSearch
     public class PersonSearchController : ControllerBase
     {
         private readonly ILogger<PersonSearchController> _logger;
-        private readonly ISearchRequestService _searchRequestService;
+        private readonly IPersonFoundService _personFoundService;
         private readonly ISearchApiRequestService _searchApiRequestService;
         private readonly IMapper _mapper;
 
-        public PersonSearchController(ISearchRequestService searchRequestService,
+        public PersonSearchController(IPersonFoundService personFoundService,
             ISearchApiRequestService searchApiRequestService, ILogger<PersonSearchController> logger,  IMapper mapper)
         {
-            _searchRequestService = searchRequestService;
+            _personFoundService = personFoundService;
             _searchApiRequestService = searchApiRequestService;
             _logger = logger;
             _mapper = mapper;
@@ -64,12 +64,7 @@ namespace DynamicsAdapter.Web.PersonSearch
                     SearchRequestId = searchRequestId
                 };
 
-            
-                await SavePerson(searchRequest, personCompletedEvent, cts.Token);
-                await UploadIdentifiers(searchRequest, personCompletedEvent, cts.Token);
-                await UploadAddresses(searchRequest, personCompletedEvent, cts.Token);
-                await UploadPhoneNumbers(searchRequest, personCompletedEvent, cts.Token);
-                await UploadNames(searchRequest, personCompletedEvent, cts.Token);
+                _personFoundService.ProcessPersonFound(personCompletedEvent.MatchedPerson, personCompletedEvent.ProviderProfile, searchRequest, cts.Token);
             }
             catch (Exception ex)
             {
@@ -169,66 +164,6 @@ namespace DynamicsAdapter.Web.PersonSearch
             return Ok();
         }
 
-        private async Task<bool> UploadIdentifiers(SSG_SearchRequest request, PersonSearchCompleted personCompletedEvent, CancellationToken concellationToken)
-        {
-            if (personCompletedEvent.MatchedPerson.Identifiers == null) return true;
-            foreach (var matchFoundPersonId in personCompletedEvent.MatchedPerson.Identifiers)
-            {
-                SSG_Identifier identifier = _mapper.Map<SSG_Identifier>(matchFoundPersonId);
-                identifier.SSG_SearchRequest = request;
-                identifier.InformationSource = personCompletedEvent.ProviderProfile.DynamicsID();
-                var identifer = await _searchRequestService.CreateIdentifier(identifier, concellationToken);
-            }
-            return true;
-        }
-
-        private async Task<bool> SavePerson(SSG_SearchRequest request, PersonSearchCompleted personCompletedEvent, CancellationToken concellationToken)
-        {
-            if (personCompletedEvent.MatchedPerson == null) return true;
-            SSG_Person person = _mapper.Map<SSG_Person>(personCompletedEvent.MatchedPerson);
-            person.SearchRequest = request;
-            person.InformationSource = personCompletedEvent.ProviderProfile.DynamicsID();
-            await _searchRequestService.SavePerson(person, concellationToken);
-            return true;
-        }
-
-        private async Task<bool> UploadAddresses(SSG_SearchRequest request, PersonSearchCompleted personCompletedEvent, CancellationToken concellationToken)
-        {
-            if (personCompletedEvent.MatchedPerson.Addresses == null) return true;
-            foreach (var address in personCompletedEvent.MatchedPerson.Addresses)
-            {
-                SSG_Address addr = _mapper.Map<SSG_Address>(address);
-                addr.SearchRequest = request;
-                addr.InformationSource = personCompletedEvent.ProviderProfile.DynamicsID();
-                var uploadedAddr = await _searchRequestService.CreateAddress(addr, concellationToken);
-            }
-            return true;
-        }
-
-        private async Task<bool> UploadPhoneNumbers(SSG_SearchRequest request, PersonSearchCompleted personCompletedEvent, CancellationToken concellationToken)
-        {
-            if (personCompletedEvent.MatchedPerson.Phones == null) return true;
-            foreach (var phone in personCompletedEvent.MatchedPerson.Phones)
-            {
-                SSG_PhoneNumber ph = _mapper.Map<SSG_PhoneNumber>(phone);
-                ph.SearchRequest = request;
-                ph.InformationSource = personCompletedEvent.ProviderProfile.DynamicsID();
-                await _searchRequestService.CreatePhoneNumber(ph, concellationToken);
-            }
-            return true;
-        }
-
-        private async Task<bool> UploadNames(SSG_SearchRequest request, PersonSearchCompleted personCompletedEvent, CancellationToken concellationToken)
-        {
-            if (personCompletedEvent.MatchedPerson.Names == null) return true;
-            foreach (var name in personCompletedEvent.MatchedPerson.Names)
-            {
-                SSG_Aliase n = _mapper.Map<SSG_Aliase>(name);
-                n.SearchRequest = request;
-                n.InformationSource = personCompletedEvent.ProviderProfile.DynamicsID();
-                await _searchRequestService.CreateName(n, concellationToken);
-            }
-            return true;
-        }
+ 
     }
 }
